Changing UTS counter_t typedef by unsigned long long; changing bots_number_of_tasks variable typefrom int to unsiged long long; changing UTS messages to get more information; including uts main function header declaration; changing some input file in UTS benchmark; close #113

+++ common/bots.h
@@ -47,7 +47,9 @@ extern char bots_ldflags[];
 /* time variables */
 extern double bots_time_program;
 extern double bots_time_sequential;
-extern int    bots_number_of_tasks;
+
+/* number of tasks variable */
+extern unsigned long long bots_number_of_tasks; /* forcing 8 bytes size on -m32 and -m64 */
 
 extern char bots_cutoff[];
 extern int  bots_cutoff_value;
